requirements-dev.txt
-r requirements.txt
black==22.3.0
isort==5.10.1
flake8==4.0.1
flake8-isort==4.1.1
pre-commit==2.19.0

django-debug-toolbar==3.4.0
django-extensions==3.1.5
ipython==8.3.0
